pyproject.toml
[tool.poetry]
name = "testchecker"
version = "0.1.0"
description = ""
authors = ["Aleksey Eliseev <alekseyeliseev@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
gitpython = "^3.1.44"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
